Moved DefinedNamesExtensions to own file
